edubackend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-multipart
pydantic
python-dotenv